--- fs-module/index.js
@@ -1,4 +1,0 @@
-import listFiles from './listFiles.js'
-import up from './up.js'
-
-export { listFiles, up }
