+++ fs/index.js
@@ -1,0 +1,4 @@
+import listFiles from './modules/listFiles.js'
+import up from './modules/up.js'
+
+export { listFiles, up }
