+++ fs/modules/listFiles.js
@@ -1,6 +1,6 @@
 import { EOL } from 'os'
 import { readdir } from 'fs/promises'
-import { printMessage, printCurrentDir } from '../msg-module/index.js'
+import { printMessage, printCurrentDir } from '../../msg/index.js'
 
 const listFiles = async () => {
   const filesArr = await readdir(process.env.CURRENT_DIR)
