+++ fs/modules/up.js
@@ -1,5 +1,5 @@
 import { join } from 'path'
-import { printCurrentDir } from '../msg-module/index.js'
+import { printCurrentDir } from '../../msg/index.js'
 
 const up = () => {
   process.env.CURRENT_DIR = join(process.env.CURRENT_DIR, '..')
