+++ index.js
@@ -1,10 +1,12 @@
 import { homedir } from 'os'
-import { closeStdin } from './utils/index.js'
-import { printStartMsg, printExitMsg } from './msg-module/index.js'
+import { printStartMsg, printExitMsg } from './msg/index.js'
 import validateByOperationRouter from './operationRouter.js'
 
-function runCli () {
-  process.env.CURRENT_DIR = homedir
+const closeStdin = () => process.stdin.destroy()
+const initEnv = () => process.env.CURRENT_DIR = homedir
+
+const runCli = () => {
+  initEnv()
   printStartMsg()
 
   process.stdin
