--- msg-module/index.js
@@ -1,6 +1,0 @@
-import printMessage from './printMessage.js'
-import printStartMsg from './printStartMsg.js'
-import printCurrentDir from './printCurrentDir.js'
-import printExitMsg from './printExitMsg.js'
-
-export { printMessage, printStartMsg, printExitMsg, printCurrentDir }
