+++ msg/index.js
@@ -1,0 +1,6 @@
+import printMessage from './modules/printMessage.js'
+import printStartMsg from './modules/printStartMsg.js'
+import printCurrentDir from './modules/printCurrentDir.js'
+import printExitMsg from './modules/printExitMsg.js'
+
+export { printMessage, printStartMsg, printExitMsg, printCurrentDir }
