+++ os/index.js
@@ -1,0 +1,4 @@
+import getCpus from './modules/getCpus.js'
+import getEol from './modules/getEol.js'
+
+export { getCpus, getEol }
