+++ os/modules/getEol.js
@@ -1,0 +1,5 @@
+import { EOL } from 'os'
+
+const getEol = () => printMessage(EOL)
+
+export default getEol
