--- utils/closeStdin.js
@@ -1,3 +1,0 @@
-const closeStdin = () => process.stdin.destroy()
-
-export default closeStdin
